% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractSNPs.R
\name{extract_snps_from_fasta}
\alias{extract_snps_from_fasta}
\title{extract_snps_from_fasta}
\usage{
extract_snps_from_fasta(
  aln_path,
  gap_freq = 0.15,
  maf_freq = 0.01,
  is_N_minor_allele = F,
  save_path = "out.fa",
  return_to_environment = T
)
}
\arguments{
\item{aln_path}{path to multi fasta alignment}

\item{gap_freq}{sites with 'N' allele (ambiguous/gap) frequency >gap_freq will be dropped (default = 0.15)}

\item{maf_freq}{sites with second most common allele (minor allele) frequency <maf_freq will be dropped (default = 0.01)}

\item{is_N_minor_allele}{specify whether 'N' allele (ambiguous/gap) should be considered as minor allele (default = F). Warning! Can lead to poor quality SNPs}

\item{save_path}{path to save fasta (default = out.fa). POS and SEQS files will also be created. Set to NULL to avoid saving. WARNING: Files will be overwritten without checking!}

\item{return_to_environment}{return snps R matrix back to environment (default = T)}
}
\value{
If return_to_environment=T, R character matrix with requested positions. colnames: positions, rownames: sequence names
}
\description{
Function to extract SNP positions from fasta alignment. If you require non-SNP positions, use extract_positions_from_fasta() instead.
}
\examples{
\dontrun{
aln_path <- system.file("extdata", "test.fa", package = "FastaR")
snp.dat <- extract_snps_from_fasta(aln_path)
}
}
