% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractSNPs.R
\name{extract_snps_from_fasta_sparse}
\alias{extract_snps_from_fasta_sparse}
\title{extract_snps_from_fasta_sparse}
\usage{
extract_snps_from_fasta_sparse(
  aln_path,
  gap_freq = 0.15,
  maf_freq = 0.01,
  is_N_minor_allele = F
)
}
\arguments{
\item{aln_path}{path to multi fasta alignment}

\item{gap_freq}{sites with 'N' allele (ambiguous/gap) frequency >gap_freq will be dropped (default = 0.15)}

\item{maf_freq}{sites with second most common allele (minor allele) frequency <maf_freq will be dropped (default = 0.01)}

\item{is_N_minor_allele}{specify whether 'N' allele (ambiguous/gap) should be considered as minor allele (default = F). Warning! Can lead to poor quality SNPs}
}
\value{
R list with SNPs in sparse format and additional parameters for fast computations
}
\description{
Function to extract SNPs from a multi-fasta alignment and return to R in sparse matrix format. This format to perform fast computations such as cross products,
see https://github.com/Sudaraka88/LDWeaver/blob/main/R/BacGWES.R for a use case.
}
\examples{
\dontrun{
aln_path <- system.file("extdata", "test.fa", package = "FastaR")
snp.dat <- extract_snps_from_fasta_sparse(aln_path)
}
}
