% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa_utils.R
\name{fa_percent_GC}
\alias{fa_percent_GC}
\title{Get GC percent for multi-fasta file}
\usage{
fa_percent_GC(fasta_file)
}
\arguments{
\item{fasta_file}{Either a path or a connection to multi-fasta file. The
input sequence file should have extention .fa or .fasta
In the sequence header: only string before first space and/or first colon (:) will be considered for futher processes.
**Important consideration when header have big names.}
}
\value{
A tibble of GC percent and a barplot of GC percent distribution
 \code{if the fasta file contains sequences less than 20}
}
\description{
for given multi-fasta file get GC percent for each sequence
}
\examples{
\dontrun{
 ref_fasta <- system.file("exdata", "Sc_nucl_R64-2-1.fasta", package = "fastaR")
 fastaR::fa_percent_GC(fasta_file=ref_fasta)
}
}
