% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa_utils.R
\name{fa_size}
\alias{fa_size}
\title{Get size of fasta sequence}
\usage{
fa_size(fasta_file)
}
\arguments{
\item{fasta_file}{Either a path or a connection to multi-fasta file.
The input sequence file should have extention .fa or .fasta}
}
\value{
Tab delimited .size file
}
\description{
Get size (in bp) for each sequence in multi-fasta file.
}
\examples{
\dontrun{

 fa_size(fasta_file="Sc_orf_coding_R64-2-1.fasta")

}
}
