% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa_utils.R
\name{fa_some_records}
\alias{fa_some_records}
\title{Get fasta sequences for a gene list.}
\usage{
fa_some_records(gene_list, fasta_file, outfile = "stdout.fa")
}
\arguments{
\item{gene_list}{A charcater vector of gene names; present in multi-fasta
file}

\item{fasta_file}{Either a path or a connection to reference multi-fasta
file, from which subset of sequences for given input list is to be
retrieved.}

\item{outfile}{A character vector containing the path to the file to write
output}
}
\value{
A multi-fasta file, containing sequences of given input list
}
\description{
Get fasta sequences of given input list from reference
  multi-fasta file.
}
\examples{
\dontrun{

 myGenelist <- scan("Sc_myGenelist.txt",  what="character", sep=NULL)

 fa_some_records(gene_list=myGenelist, fasta_file="Sc_orf_trans_all_R64-2-1.fasta", outfile="sc_myGenelist.fa")

}
}
