% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa_utils.R
\name{faPercentGC}
\alias{faPercentGC}
\title{Get GC percent}
\usage{
faPercentGC(fasta_file)
}
\arguments{
\item{fasta_file}{Either a path or a connection to multi-fasta file. The
input sequence file should have extention .fa or .fasta}
}
\value{
A output file of GC percent and a barplot of GC percent distribution
 \code{if the fasta file contains sequences less than 20}
}
\description{
for given multi-fasta file get GC percent for each sequence
}
\examples{
\dontrun{

 faPercentGC(fasta_file="Sc_orf_coding_R64-2-1.fasta")
}
}
