% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fasta_utils.R
\name{get_fasta_from_bed}
\alias{get_fasta_from_bed}
\title{get fasta sequences for bed file}
\usage{
get_fasta_from_bed(bedFile, fasta_file, outfile = "out")
}
\arguments{
\item{bedFile}{A character vector containing path to the file to the bed
file, file format should be standard UCSC bed format with column:
\code{ 1. chromosome-id, 2. start, 3. end, 4. name, 5. score, 6. strand}}

\item{fasta_file}{Either a path or a connection to reference multi-fasta
file, from which subset of sequences for given input feature is to be
retrieved.
In the sequence header: only string before first space and/or first colon (:)
will be considered for futher processes.
**Important consideration when header have big names.}

\item{outfile}{A character vector, determining outfile name, default: out}
}
\value{
A output fasta file of sequences in given bed region
}
\description{
Extracts sequences from a fasta file for each of the intervals defined in a BED file
}
\examples{
\dontrun{

get_fasta_from_bed(bedFile="Sc_orf_coding_R64-2-1.bed",
fasta_file="S288C_reference_sequence_R64-2-1_20150113.fa",
outfile="Sc_subset")

}
}
