% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractSNPs.R
\name{write_fasta}
\alias{write_fasta}
\title{write_fasta}
\usage{
write_fasta(snps, fa_path)
}
\arguments{
\item{snps}{snps in matrix format}

\item{fa_path}{path to save fasta}
}
\description{
Function to write a sequence matrix to fasta
}
\examples{
\dontrun{
aln_path <- write_fasta(snps, "snps.fa")
}
}
