% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa_utils.R
\name{fa_summary}
\alias{fa_summary}
\title{Fasta file summary}
\usage{
fa_summary(fasta_file)
}
\arguments{
\item{fasta_file}{Either a path or a connection to multi-fasta file.
The input sequence file should have extention .fa or .fasta
In the sequence header: only string before first space and/or first colon (:) will be considered for futher processes.
**Important consideration when header have big names.}
}
\value{
A summary table
}
\description{
Get summary of input multi-fasta file like mean length, median
  length, GC content etc
}
\examples{
\dontrun{

 fa_summary(fasta_file="Sc_orf_coding_R64-2-1.fasta")
}
}
