% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fasta_utils.R
\name{get_random_sequences_from_fasta}
\alias{get_random_sequences_from_fasta}
\title{get_random_sequences_from_fasta}
\usage{
get_random_sequences_from_fasta(
  fasta_file,
  numberOfRandomSequences = 50,
  lengthOfRandomSequence = 100,
  outfile = "randomSeq"
)
}
\arguments{
\item{fasta_file}{Either a path or a connection to reference multi-fasta
file, from which random sequences need to be generated. In the sequence header: only string before first space and/or
first colon (:) will be considered for futher processes. **Important
consideration when header have big names.}

\item{numberOfRandomSequences}{Numeric, number of random sequences to be generated.}

\item{lengthOfRandomSequence}{Numeric, length of random sequences.}

\item{outfile}{A character vector, determining outfile name, \code{default:
randomSeq }}
}
\value{
A bedfile of random regions and fasta file containing random sequences.
}
\description{
Generate n-random sequences of pre-determined length from given input multi-fasta file
}
\examples{
\dontrun{

fasta_file <- "S288C_reference_sequence_R64-2-1_20150113.fa"

get_random_sequences_from_fasta(fasta_file = fasta_file, numberOfRandomSequences = 100, lengthOfRandomSequence = 500, outfile = "Sc_randomSeq")

}

}
